Fixed Fines, added more constraints all of my forms, changed the script
The fines completely work, and were moved out of PatronTab and into a
separate file so fines aren't added every time you load patron tab.
Placed contraints on adding and editing patrons, and added constraints
for paying fines. I also set up a the search on the main page. I added a
unique auto updating key to loans and removed amountCharged as I was not
using it and it seem redundant.